set type WiFi
set WiFi  STA -> softAP
set ledPin D4 -> D6
add initServer (from initWiFi)
