+++ Smart-LED/include/config.h
@@ -3,8 +3,7 @@
 
 #include <Arduino.h>
 
-inline const uint8_t ledPin = D4;
-inline const uint8_t buttonPin = D5; //
+inline const uint8_t ledPin = D6; 
 
 inline const uint8_t NUMBER_OF_LEDS = 30; // 60
 
@@ -12,7 +11,7 @@ inline const uint32_t DEBOUNCE_TIME = 500000; // ms //
 
 inline const uint8_t EEPROM_SIZE = 16;
 
-inline const char* ssid = "Test";
-inline const char* password = "123123123";
+inline const char* ssidESP = "Smart-LED (192.168.0.100)";
+inline const char* passwordESP = "123123123";
 
 #endif //CONFIG_H
