+++ Smart-LED/include/initServer.h
@@ -1,0 +1,6 @@
+#ifndef INIT_SERVER_H
+#define INIT_SERVER_H
+
+void initServer();
+
+#endif // INIT_SERVER_H
