+++ Smart-LED/src/main.cpp
@@ -11,7 +11,6 @@ Color colorLeds;
 
 bool flagSetLeds = true;
 bool flagStopLeds = false;
-bool flagNotifyClients = false;
 
 void setup()
 {
@@ -22,13 +21,7 @@ void setup()
 }
 
 void loop()
-{
-    if (flagNotifyClients == true)
-    {
-        notifyClients();
-        flagNotifyClients = false;
-    }
-    
+{   
     if (goLighting == true)
     {
         if (flagSetLeds == true)
